+++ extension/src/libsm64_mario_state.cpp
@@ -5,9 +5,12 @@
 LibSM64MarioState::LibSM64MarioState(const SM64MarioState &state, godot::real_t scale_factor) :
 		position(-state.position[2] / scale_factor, state.position[1] / scale_factor, state.position[0] / scale_factor),
 		velocity(-state.velocity[2] / (scale_factor * LibSM64::tick_delta_time), state.velocity[1] / (scale_factor * LibSM64::tick_delta_time), state.velocity[0] / (scale_factor * LibSM64::tick_delta_time)),
+		forward_velocity(state.forwardVelocity / (scale_factor * LibSM64::tick_delta_time)),
 		face_angle(state.faceAngle),
 		health(state.health),
 		action(state.action),
+		anim_id(static_cast<LibSM64::MarioAnimID>(state.animID)),
+		anim_frame(state.animFrame),
 		flags(state.flags),
 		particle_flags(state.particleFlags),
 		invincibility_time(state.invincTimer * LibSM64::tick_delta_time) {}
@@ -26,6 +29,14 @@ void LibSM64MarioState::set_velocity(const godot::Vector3 &p_value) {
 
 godot::Vector3 LibSM64MarioState::get_velocity() const {
 	return velocity;
+}
+
+void LibSM64MarioState::set_forward_velocity(float p_value) {
+	forward_velocity = p_value;
+}
+
+float LibSM64MarioState::get_forward_velocity() const {
+	return forward_velocity;
 }
 
 void LibSM64MarioState::set_face_angle(float p_value) {
@@ -50,6 +61,22 @@ void LibSM64MarioState::set_action(godot::BitField<LibSM64::ActionFlags> p_value
 
 godot::BitField<LibSM64::ActionFlags> LibSM64MarioState::get_action() const {
 	return action;
+}
+
+void LibSM64MarioState::set_anim_id(LibSM64::MarioAnimID p_value) {
+	anim_id = p_value;
+}
+
+LibSM64::MarioAnimID LibSM64MarioState::get_anim_id() const {
+	return anim_id;
+}
+
+void LibSM64MarioState::set_anim_frame(int p_value) {
+	anim_frame = p_value;
+}
+
+int LibSM64MarioState::get_anim_frame() const {
+	return anim_frame;
 }
 
 void LibSM64MarioState::set_flags(godot::BitField<LibSM64::MarioFlags> p_value) {
@@ -83,6 +110,9 @@ void LibSM64MarioState::_bind_methods() {
 	godot::ClassDB::bind_method(godot::D_METHOD("set_velocity", "value"), &LibSM64MarioState::set_velocity);
 	godot::ClassDB::bind_method(godot::D_METHOD("get_velocity"), &LibSM64MarioState::get_velocity);
 	ADD_PROPERTY(godot::PropertyInfo(godot::Variant::VECTOR3, "velocity"), "set_velocity", "get_velocity");
+	godot::ClassDB::bind_method(godot::D_METHOD("set_forward_velocity", "value"), &LibSM64MarioState::set_forward_velocity);
+	godot::ClassDB::bind_method(godot::D_METHOD("get_forward_velocity"), &LibSM64MarioState::get_forward_velocity);
+	ADD_PROPERTY(godot::PropertyInfo(godot::Variant::FLOAT, "forward_velocity"), "set_forward_velocity", "get_forward_velocity");
 	godot::ClassDB::bind_method(godot::D_METHOD("set_face_angle", "value"), &LibSM64MarioState::set_face_angle);
 	godot::ClassDB::bind_method(godot::D_METHOD("get_face_angle"), &LibSM64MarioState::get_face_angle);
 	ADD_PROPERTY(godot::PropertyInfo(godot::Variant::FLOAT, "face_angle"), "set_face_angle", "get_face_angle");
@@ -92,6 +122,12 @@ void LibSM64MarioState::_bind_methods() {
 	godot::ClassDB::bind_method(godot::D_METHOD("set_action", "value"), &LibSM64MarioState::set_action);
 	godot::ClassDB::bind_method(godot::D_METHOD("get_action"), &LibSM64MarioState::get_action);
 	ADD_PROPERTY(godot::PropertyInfo(godot::Variant::INT, "action"), "set_action", "get_action");
+	godot::ClassDB::bind_method(godot::D_METHOD("set_anim_id", "value"), &LibSM64MarioState::set_anim_id);
+	godot::ClassDB::bind_method(godot::D_METHOD("get_anim_id"), &LibSM64MarioState::get_anim_id);
+	ADD_PROPERTY(godot::PropertyInfo(godot::Variant::INT, "anim_id"), "set_anim_id", "get_anim_id");
+	godot::ClassDB::bind_method(godot::D_METHOD("set_anim_frame", "value"), &LibSM64MarioState::set_anim_frame);
+	godot::ClassDB::bind_method(godot::D_METHOD("get_anim_frame"), &LibSM64MarioState::get_anim_frame);
+	ADD_PROPERTY(godot::PropertyInfo(godot::Variant::INT, "anim_frame"), "set_anim_frame", "get_anim_frame");
 	godot::ClassDB::bind_method(godot::D_METHOD("set_flags", "value"), &LibSM64MarioState::set_flags);
 	godot::ClassDB::bind_method(godot::D_METHOD("get_flags"), &LibSM64MarioState::get_flags);
 	ADD_PROPERTY(godot::PropertyInfo(godot::Variant::INT, "flags"), "set_flags", "get_flags");